+++ Patches/TESTING.cs
@@ -1,4 +1,4 @@
-﻿/*using BepInEx.Logging;
+﻿using BepInEx.Logging;
 using GameNetcodeStuff;
 using HarmonyLib;
 using System;
@@ -8,6 +8,7 @@ using Unity.Collections.LowLevel.Unsafe;
 using Unity.Netcode;
 using UnityEngine;
 using UnityEngine.UIElements;
+using static SCP956.Plugin;
 
 namespace SCP956.Patches
 {
@@ -15,14 +16,6 @@ namespace SCP956.Patches
     internal class TESTING : MonoBehaviour
     {
         private static ManualLogSource logger = Plugin.LoggerInstance;
-
-        private static PlayerControllerB localPlayer { get { return StartOfRound.Instance.localPlayerController; } }
-
-        static bool invis = false;
-        static bool noSpawning = false;
-        static bool godMode = false;
-
-        static List<string> commands = new List<string> { "/spawn", "/spawnitem", "/invis", "/teleport", "/nospawning", "/money", "/godmode", "/damage", "/refresh", "/lights", "/enemies", "/players", "/items", "/turret" };
 
         [HarmonyPostfix, HarmonyPatch(typeof(HUDManager), nameof(HUDManager.PingScan_performed))]
         public static void PingScan_performedPostFix()
@@ -39,179 +32,30 @@ namespace SCP956.Patches
 
 
             // Comment these out
-            if (args[0] == "/help")
+            if (args[0] == "/setVoice") // TODO: Test this
             {
-                foreach (var command in commands)
+                float num = StartOfRound.Instance.drunknessSideEffect.Evaluate(float.Parse(args[1]));
+                if (num > 0.15f)
                 {
-                    logger.LogDebug(command);
+                    SoundManager.Instance.playerVoicePitchTargets[localPlayer.actualClientId] = 1f + num;
                 }
-            }
-            if (args[0] == "/spawn")
-            {
-                Vector3 pos = localPlayer.transform.forward * 2f + localPlayer.transform.position;
-                int index = RoundManager.Instance.currentLevel.Enemies.FindIndex(x => x.enemyType.enemyName == args[1]);
-                RoundManager.Instance.SpawnEnemyOnServer(pos, UnityEngine.Random.Range(0f, 360f), index);
-                HUDManager.Instance.DisplayTip("Testing", $"Spawned enemy: {args[1]}");
-                RoundManager.Instance.RefreshEnemiesList();
-            }
-            if (args[0] == "/spawnitem")
-            {
-                args[1].Replace("_", " ");
-                Vector3 pos = localPlayer.transform.forward * 2f + localPlayer.transform.position;
-                Item item = StartOfRound.Instance.allItemsList.itemsList.Where(x => x.itemName == args[1]).FirstOrDefault();
-                if (item == null) { return; }
-
-                GameObject obj = UnityEngine.Object.Instantiate(item.spawnPrefab, pos, Quaternion.identity, StartOfRound.Instance.propsContainer);
-                obj.GetComponent<NetworkObject>().Spawn();
-                HUDManager.Instance.DisplayTip("Testing", $"Spawned item: {item.itemName}");
-            }
-            if (args[0] == "/invis")
-            {
-                if (args[1] == "true")
+                else
                 {
-                    invis = true;
+                    SoundManager.Instance.playerVoicePitchTargets[localPlayer.actualClientId] = 1f;
                 }
-                else if (args[1] == "false")
-                {
-                    invis = false;
-                }
-                HUDManager.Instance.DisplayTip("Testing", $"Invis set to {invis}");
-            }
-            if (args[0] == "/teleport")
-            {
-                if (args[1] == "main")
-                {
-                    Vector3 pos = RoundManager.FindMainEntrancePosition(true, true);
-                    localPlayer.TeleportPlayer(pos);
-                }
-            }
-            if (args[0] == "/nospawning")
-            {
-                if (args[1] == "true")
-                {
-                    noSpawning = true;
-                }
-                else if (args[1] == "false")
-                {
-                    noSpawning = false;
-                }
-                HUDManager.Instance.DisplayTip("Testing", $"Enemy spawning set to {!noSpawning}");
-            }
-            if (args[0] == "/money")
-            {
-                int amount = Convert.ToInt32(args[1]);
-                FindObjectOfType<Terminal>().groupCredits += amount;
-                HUDManager.Instance.DisplayTip("Testing", $"Credits added: {amount}");
-            }
-            if (args[0] == "/godmode")
-            {
-                if (args[1] == "true")
-                {
-                    godMode = true;
-                }
-                else if (args[1] == "false")
-                {
-                    godMode = false;
-                }
-                HUDManager.Instance.DisplayTip("Testing", $"Godmode set to {godMode}");
-            }
-            if (args[0] == "/damage")
-            {
-                int amount = Convert.ToInt32(args[1]);
-                localPlayer.DamagePlayer(amount);
-                HUDManager.Instance.DisplayTip("Testing", $"Damaged self: {amount}");
-            }
-            if (args[0] == "/refresh")
-            {
-                RoundManager.Instance.RefreshEnemiesList();
-                RoundManager.Instance.RefreshEnemyVents();
-                RoundManager.Instance.RefreshLightsList();
-            }
-            if (args[0] == "/lights")
-            {
-                bool value = Convert.ToBoolean(args[1]);
-                RoundManager.Instance.TurnOnAllLights(value);
-                HUDManager.Instance.DisplayTip("Testing", $"Lights set to {value}");
-            }
-            if (msg == "/enemies")
-            {
-                logger.LogDebug("ENEMIES:");
-                foreach (var enemy in GetEnemies())
-                {
-                    logger.LogDebug(enemy.enemyType.enemyName);
-                }
-            }
-            if (msg == "/players")
-            {
-                logger.LogDebug("PLAYERS:");
-                foreach (var player in StartOfRound.Instance.allPlayerScripts)
-                {
-                    logger.LogDebug(player.playerUsername);
-                }
-            }
-            if (msg == "/items")
-            {
-                logger.LogDebug("ITEMS:");
-                foreach (var item in StartOfRound.Instance.allItemsList.itemsList)
-                {
-                    logger.LogDebug(item.itemName);
-                }
-            }
-            if (msg == "/turret")
-            {
-                Vector3 pos = localPlayer.transform.forward * 2f + localPlayer.transform.position;
-                GameObject turretObj = UnityEngine.Object.Instantiate(StartOfRound.Instance.levels.Where(x => x.PlanetName == "41 Experimentation").First().spawnableMapObjects.Where(x => x.prefabToSpawn.name == "TurretContainer").First().prefabToSpawn, pos, Quaternion.identity, RoundManager.Instance.mapPropsContainer.transform);
-                turretObj.GetComponent<NetworkObject>().Spawn(true);
-                HUDManager.Instance.DisplayTip("Testing", "Spawned Turret");
             }
         }
-
-        [HarmonyPrefix, HarmonyPatch(typeof(EnemyAI), nameof(EnemyAI.PlayerIsTargetable))]
-        public static bool PlayerIsTargetablePrefix(ref bool __result)
-        {
-            if (NetworkManager.Singleton.IsServer || NetworkManager.Singleton.IsHost)
-            {
-                if (invis)
-                {
-                    __result = false;
-                    return false;
-                }
-            }
-            return true;
-        }
-
-        [HarmonyPrefix, HarmonyPatch(typeof(RoundManager), nameof(RoundManager.SpawnEnemiesOutside))]
-        public static bool SpawnEnemiesOutsidePrefix()
-        {
-
-            if (noSpawning)
-            {
-                return false;
-            }
-            return true;
-        }
-
-        [HarmonyPrefix, HarmonyPatch(typeof(RoundManager), nameof(RoundManager.SpawnInsideEnemiesFromVentsIfReady))]
-        public static bool SpawnInsideEnemiesFromVentsIfReadyPrefix()
-        {
-
-            if (noSpawning)
-            {
-                return false;
-            }
-            return true;
-        }
-
-        [HarmonyPrefix, HarmonyPatch(typeof(PlayerControllerB), nameof(PlayerControllerB.DamagePlayer))]
-        public static bool DamagePlayerPrefix(int damageNumber)
-        {
-            if (godMode)
-            {
-                HUDManager.Instance.DisplayTip("Testing", $"Godmode prevented {damageNumber} damage to player");
-                return false;
-            }
-            return true;
-        }
+        /*
+         float num11 = StartOfRound.Instance.drunknessSideEffect.Evaluate(drunkness);
+if (num11 > 0.15f)
+{
+    SoundManager.Instance.playerVoicePitchTargets[playerClientId] = 1f + num11;
+}
+else
+{
+    SoundManager.Instance.playerVoicePitchTargets[playerClientId] = 1f;
+}
+         */
 
         public static List<SpawnableEnemyWithRarity> GetEnemies()
         {
@@ -229,4 +73,4 @@ namespace SCP956.Patches
             return enemies;
         }
     }
-}*/
+}